feat: Add option for custom render logic in controllers (Fixes #4)
If a controller extends from a JavaFX Parent the instance itself will be used as the rendered view.

If the specified view starts with a '#' it will be treated as a method reference. If the given methods exists within the class and returns a Parent, it will be used as the rendered view.