feat(framework): Lower requirements for components
Controllers can now use a Node as their view to be more inline with JavaFX.
When displaying a controller/component as a root of a scene (e.g. using show/modals), it still has to provide a Parent as its view as the root of a scene has to be a Parent.

TODO: More type checking and edge cases (e.g. in the processor)